+++ apps/api/src/app/admin/update-asset-profile.dto.ts
@@ -1,6 +1,6 @@
 import { IsCurrencyCode } from '@ghostfolio/api/validators/is-currency-code';
 
-import { AssetClass, AssetSubClass, Prisma } from '@prisma/client';
+import { AssetClass, AssetSubClass, DataSource, Prisma } from '@prisma/client';
 import {
   IsArray,
   IsEnum,
@@ -19,8 +19,8 @@ export class UpdateAssetProfileDto {
   @IsOptional()
   assetSubClass?: AssetSubClass;
 
-  @IsString()
   @IsOptional()
+  @IsString()
   comment?: string;
 
   @IsArray()
@@ -31,8 +31,12 @@ export class UpdateAssetProfileDto {
   @IsOptional()
   currency?: string;
 
-  @IsString()
+  @IsEnum(DataSource, { each: true })
   @IsOptional()
+  dataSource?: DataSource;
+
+  @IsOptional()
+  @IsString()
   name?: string;
 
   @IsObject()
@@ -42,6 +46,10 @@ export class UpdateAssetProfileDto {
   @IsArray()
   @IsOptional()
   sectors?: Prisma.InputJsonArray;
+
+  @IsOptional()
+  @IsString()
+  symbol?: string;
 
   @IsObject()
   @IsOptional()