+++ apps/client/src/app/pages/about/terms-of-service/terms-of-service-page.component.ts
@@ -2,12 +2,12 @@ import { Component, OnDestroy } from '@angular/core';
 import { Subject } from 'rxjs';
 
 @Component({
-  selector: 'gf-privacy-policy-page',
+  selector: 'gf-terms-of-service-page',
   standalone: false,
-  styleUrls: ['./privacy-policy-page.scss'],
-  templateUrl: './privacy-policy-page.html'
+  styleUrls: ['./terms-of-service-page.scss'],
+  templateUrl: './terms-of-service-page.html'
 })
-export class PrivacyPolicyPageComponent implements OnDestroy {
+export class TermsOfServicePageComponent implements OnDestroy {
   private unsubscribeSubject = new Subject<void>();
 
   public ngOnDestroy() {